% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-color.R
\name{checkDim}
\alias{checkDim}
\title{checkDim}
\usage{
checkDim(x, data)
}
\arguments{
\item{x}{The color vector}

\item{data}{The dataset}
}
\description{
Check the length of the color vector. It must match the
  number of rows of the dataset.
}
