% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-popup.R
\name{checkDimPop}
\alias{checkDimPop}
\title{checkDim}
\usage{
checkDimPop(x, data)
}
\arguments{
\item{x}{The popup vector}

\item{data}{The dataset}
}
\description{
Check the length of the popup vector. It must match the
  number of rows of the dataset.
}
