% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafgl_package.R
\docType{package}
\name{leafgl-package}
\alias{leafgl}
\alias{leafgl-package}
\title{leafgl: High-Performance 'WebGl' Rendering for Package 'leaflet'}
\description{
Provides bindings to the 'Leaflet.glify' JavaScript library which extends the 'leaflet' JavaScript library to render large data in the browser using 'WebGl'.
}
\author{
\strong{Maintainer}: Tim Appelhans \email{tim.appelhans@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Colin Fay \email{colin@thinkr.fr} (\href{https://orcid.org/0000-0001-7343-1846}{ORCID}) [contributor]
  \item Robert Plummer (Leaflet.glify plugin) [contributor]
  \item Kent Johnson \email{kjohnson@akoyabio.com} [contributor]
  \item Sebastian Gatscha \email{sebastian_gatscha@gmx.at} [contributor]
}

}
\keyword{internal}
