% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utils.R
\name{make_color_matrix}
\alias{make_color_matrix}
\title{make_color_matrix}
\usage{
make_color_matrix(x, data, alpha, palette, ...)
}
\arguments{
\item{x}{Object representing the color. Can be of class integer, numeric, Date,
POSIX*, character with color names or HEX codes, factor, matrix, data.frame,
list, json or formula.}

\item{data}{The dataset}

\item{alpha}{Should alpha be included. Default is FALSE}

\item{palette}{Name of a color palette. If \code{colourvalues} is installed, it is
passed to \code{\link[colourvalues]{colour_values_rgb}}. To see all available
palettes, please use \code{\link[colourvalues]{colour_palettes}}.
If \code{colourvalues} is not installed, the palette is passed to
\code{\link[leaflet]{colorNumeric}}.}

\item{...}{Passed to \code{\link[colourvalues]{colour_palettes}} or
\code{\link[leaflet]{colorNumeric}}.}
}
\description{
Transform object to rgb color matrix
}
\examples{
{
## For Integer/Numeric/Factor
make_color_matrix(23L)
make_color_matrix(23)
make_color_matrix(as.factor(23))

## For POSIXt / Date
make_color_matrix(as.POSIXlt(Sys.time(), "America/New_York"), NULL)
make_color_matrix(Sys.time(), NULL)
make_color_matrix(Sys.Date(), NULL)

## For matrix/data.frame
make_color_matrix(cbind(130,1,1), NULL)
make_color_matrix(matrix(1:99, ncol = 3, byrow = TRUE), data.frame(x=c(1:33)))
make_color_matrix(data.frame(matrix(1:99, ncol = 3, byrow = TRUE)), data.frame(x=c(1:33)))

## For characters
library(leaflet)
make_color_matrix("red", breweries91)
make_color_matrix("blue", breweries91)
make_color_matrix("#36ba01", breweries91)
make_color_matrix("founded", data.frame(breweries91))

## For formulaes
make_color_matrix(~founded, breweries91)
make_color_matrix(~founded + zipcode, breweries91)

## For JSON
library(jsonify)
make_color_matrix(jsonify::to_json(data.frame(r = 54, g = 186, b = 1)), NULL)

## For Lists
make_color_matrix(list(1,2), data.frame(x=c(1,2)))
}
}
