% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-popup.R
\name{make_popup}
\alias{make_popup}
\title{make_popup}
\usage{
make_popup(x, data)
}
\arguments{
\item{x}{Object representing the popup}

\item{data}{The dataset}
}
\description{
Transform object to popup
}
