% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glify-remove-clear.R
\name{remove}
\alias{remove}
\alias{removeGlPoints}
\alias{removeGlPolylines}
\alias{removeGlPolygons}
\alias{clearGlLayers}
\alias{clearGlGroup}
\title{Remove Leaflet.Glify elements from a map}
\usage{
removeGlPoints(map, layerId)

removeGlPolylines(map, layerId)

removeGlPolygons(map, layerId)

clearGlLayers(map)

clearGlGroup(map, group)
}
\arguments{
\item{map}{a map widget object, possibly created from \code{\link[leaflet:leaflet]{leaflet()}}
but more likely from \code{\link[leaflet:leafletProxy]{leafletProxy()}}}

\item{layerId}{character vector; the layer id(s) of the item to remove}

\item{group}{the name of the group whose members should be removed}
}
\value{
the new `map` object
}
\description{
Remove one or more features from a map, identified by `layerId`;
or, clear all features of the given group.
}
