% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glify-remove-clear.R
\name{removeGlPoints}
\alias{removeGlPoints}
\title{removeGlPoints}
\usage{
removeGlPoints(map, layerId)
}
\arguments{
\item{map}{The map widget}

\item{layerId}{The layerId to remove}
}
\description{
Remove points from a map, identified by layerId;
}
