% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glify-remove-clear.R
\name{removeGlPolygons}
\alias{removeGlPolygons}
\title{removeGlPolygons}
\usage{
removeGlPolygons(map, layerId)
}
\arguments{
\item{map}{The map widget}

\item{layerId}{The layerId to remove}
}
\description{
Remove polygons from a map, identified by layerId;
}
