% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glify-remove-clear.R
\name{removeGlPolylines}
\alias{removeGlPolylines}
\title{removeGlPolylines}
\usage{
removeGlPolylines(map, layerId)
}
\arguments{
\item{map}{The map widget}

\item{layerId}{The layerId to remove}
}
\description{
Remove lines from a map, identified by layerId;
}
