% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glify-shiny.R
\name{renderGlify}
\alias{renderGlify}
\title{Use Glify in shiny}
\usage{
renderGlify(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates an HTML widget}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
A server function for rendering Glify
}
\description{
Use Glify in shiny
}
\examples{
\dontrun{
library(mapview)
library(leaflet)
library(leaflet.glify)
library(sf)

n = 1e4
df1 = data.frame(id = 1:n,
    x = rnorm(n, 10, 3),
    y = rnorm(n, 49, 1.8))
pts = st_as_sf(df1, coords = c("x", "y"), crs = 4326)

m = leaflet() \%>\%
 addProviderTiles(provider = providers$CartoDB.DarkMatter) \%>\%
 addGlifyPoints(data = pts, group = "pts") \%>\%
 addMouseCoordinates() \%>\%
 setView(lng = 10.5, lat = 49.5, zoom = 6) \%>\%
 addLayersControl(overlayGroups = "pts")

ui <- fluidPage(
    glifyOutput("mymap")
)

server <- function(input, output, session) {
    output$mymap <- renderLeaflet(m)
}

shinyApp(ui, server)
}

}
